renamed 'page' components to xPage.js, moved to pages dir
