+++ BeerTier/client/src/views/ApplicationViews.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { Routes, Route, Navigate } from "react-router-dom";
-import BeerList from "../components/BeerList";
-import Hello from "../components/Hello";
+import AllBeersPage from "../pages/AllBeersPage";
+import Hello from "../pages/Hello";
 import LoginPage from "../pages/LoginPage";
 import RegisterPage from "../pages/RegisterPage";
 import BeerDetailsPage from "../pages/BeerDetailsPage";
-import BeerForm from "../pages/BeerForm";
+import NewBeerPage from "../pages/NewBeerPage";
 
 export default function ApplicationViews({ isLoggedIn }) {
   return (
@@ -15,11 +15,11 @@ export default function ApplicationViews({ isLoggedIn }) {
           <Route index element={<Hello />} />
           <Route path="/login" element={<LoginPage />} />
           <Route path="/register" element={<RegisterPage />} />
-          <Route path="beer" element={<BeerList />} />
+          <Route path="beer" element={<AllBeersPage />} />
           <Route path="beer/:id" element={<BeerDetailsPage />} />
           <Route
             path="/beer/new"
-            element={isLoggedIn ? <BeerForm /> : <Navigate to="/login" />}
+            element={isLoggedIn ? <NewBeerPage /> : <Navigate to="/login" />}
           />
           <Route
             path="/beer/me"
